Api-Data-Pipeline-Integration/Get_API_Data/requirements.txt
# Function dependencies, for example:
# package>=version
requests
google-cloud-storage
pandas